+++ src/main/kotlin/app/revanced/patches/reddit/customclients/boostforreddit/api/SpoofClientPatch.kt
@@ -6,7 +6,9 @@ import app.revanced.patcher.fingerprint.method.impl.MethodFingerprintResult
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import app.revanced.patches.reddit.customclients.AbstractSpoofClientPatch
+import app.revanced.patches.reddit.customclients.Constants.OAUTH_USER_AGENT
 import app.revanced.patches.reddit.customclients.boostforreddit.api.fingerprints.GetClientIdFingerprint
+import app.revanced.patches.reddit.customclients.boostforreddit.api.fingerprints.LoginActivityOnCreateFingerprint
 
 @Patch(
     name = "Spoof client",
@@ -15,7 +17,9 @@ import app.revanced.patches.reddit.customclients.boostforreddit.api.fingerprints
 )
 @Suppress("unused")
 object SpoofClientPatch : AbstractSpoofClientPatch(
-    "http://rubenmayayo.com", listOf(GetClientIdFingerprint)
+    "http://rubenmayayo.com",
+    clientIdFingerprints = listOf(GetClientIdFingerprint),
+    userAgentFingerprints = listOf(LoginActivityOnCreateFingerprint)
 ) {
     override fun List<MethodFingerprintResult>.patchClientId(context: BytecodeContext) {
         first().mutableMethod.addInstructions(
@@ -25,5 +29,21 @@ object SpoofClientPatch : AbstractSpoofClientPatch(
                  return-object v0
             """
         )
+    }
+
+    override fun List<MethodFingerprintResult>.patchUserAgent(context: BytecodeContext) {
+        first().let { result ->
+            result.mutableMethod.apply {
+                val insertIndex = result.scanResult.patternScanResult!!.endIndex
+
+                addInstructions(
+                    insertIndex,
+                    """
+                        const-string v7, "$OAUTH_USER_AGENT"
+                        invoke-virtual {v4, v7}, Landroid/webkit/WebSettings;->setUserAgentString(Ljava/lang/String;)V
+                    """
+                )
+            }
+        }
     }
 }