+++ src/main/kotlin/app/revanced/patches/reddit/customclients/syncforreddit/api/SpoofClientPatch.kt
@@ -10,9 +10,11 @@ import app.revanced.patcher.fingerprint.method.impl.MethodFingerprintResult
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import app.revanced.patches.reddit.customclients.AbstractSpoofClientPatch
+import app.revanced.patches.reddit.customclients.Constants.OAUTH_USER_AGENT
 import app.revanced.patches.reddit.customclients.syncforreddit.api.fingerprints.GetAuthorizationStringFingerprint
 import app.revanced.patches.reddit.customclients.syncforreddit.api.fingerprints.GetBearerTokenFingerprint
 import app.revanced.patches.reddit.customclients.syncforreddit.api.fingerprints.ImgurImageAPIFingerprint
+import app.revanced.patches.reddit.customclients.syncforreddit.api.fingerprints.LoadBrowserURLFingerprint
 import app.revanced.patches.reddit.customclients.syncforreddit.detection.piracy.DisablePiracyDetectionPatch
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
@@ -33,6 +35,7 @@ import java.util.*
 object SpoofClientPatch : AbstractSpoofClientPatch(
     "http://redditsync/auth",
     clientIdFingerprints = listOf(GetAuthorizationStringFingerprint),
+    userAgentFingerprints = listOf(LoadBrowserURLFingerprint),
     miscellaneousFingerprints = listOf(ImgurImageAPIFingerprint)
 ) {
     override fun List<MethodFingerprintResult>.patchClientId(context: BytecodeContext) {
@@ -78,5 +81,19 @@ object SpoofClientPatch : AbstractSpoofClientPatch(
             apiUrlIndex,
             "const-string v1, \"https://api.imgur.com/3/image\""
         )
+    }
+
+    override fun List<MethodFingerprintResult>.patchUserAgent(context: BytecodeContext) {
+        first().let { result ->
+            val insertIndex = result.scanResult.patternScanResult!!.startIndex
+
+            result.mutableMethod.addInstructions(
+                insertIndex,
+                """
+                    const-string v0, "$OAUTH_USER_AGENT"
+                    invoke-virtual {p1, v0}, Landroid/webkit/WebSettings;->setUserAgentString(Ljava/lang/String;)V
+                """
+            )
+        }
     }
 }